feat: update Tailwind CSS dependencies and restructure authentication components
