+++ src/main.ts
@@ -1,10 +1,9 @@
 import './assets/main.css'
-
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
 
 import App from './App.vue'
-import router from './router'
+import router from './presentation/router'
 
 const app = createApp(App)
